+++ src/main/java/com/yooyoung/clotheser/global/jwt/JwtAuthenticationFilter.java
@@ -3,9 +3,8 @@ package com.yooyoung.clotheser.global.jwt;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yooyoung.clotheser.global.entity.*;
 import com.yooyoung.clotheser.user.domain.RefreshToken;
-import com.yooyoung.clotheser.user.dto.TokenResponse;
+import com.yooyoung.clotheser.user.dto.response.TokenResponse;
 import com.yooyoung.clotheser.user.repository.RefreshTokenRepository;
-import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
