+++ src/main/java/com/yooyoung/clotheser/user/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.yooyoung.clotheser.user.dto;
+package com.yooyoung.clotheser.user.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.yooyoung.clotheser.user.domain.User;
@@ -25,7 +25,7 @@ public class SignUpRequest {
 
     @Schema(title = "이메일", description = "255자 이내", example = "noonsong@gmail.com", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotBlank(message = "이메일을 입력해주세요.")
-    @Pattern(regexp = "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$",
+    @Pattern(regexp = "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*\\.[a-zA-Z]{2,}$",
             message = "이메일 형식이 맞지 않습니다.")
     private String email;
